some fixes

+++ portal/portal.lua
@@ -15,17 +15,19 @@ function handle_request(env)
   if a ~= 0 then
     if url_path == "support" then
       support.troubleshoot()
+    elseif url_path == "autossh" then
+      if support.get_autossh_status() then
+        support.stop_autossh()
+      else
+        support.start_autossh()
+      end
+      uhttpd.send("Status: 302 Found\r\n")
+      uhttpd.send("Location: http://cloudgate.citypassenger.com\r\n")
+      uhttpd.send("Content-Type: text/html\r\n\r\n")
     else
       dofile("/portal/diagnostic.lua")
     end
-    os.exit()
-  end
-
-  -- Return a wordpress error page when wifi is scheduled to be down.
-  local b = os.execute("/usr/bin/test -e /tmp/noaccess")
-  if b == 0 then
-    portal_proxy.no_wifi()
-    os.exit()
+    return true
   end
 
   local user_ip = env.REMOTE_ADDR
@@ -39,44 +41,42 @@ function handle_request(env)
     end
   end
   f:close()
-
-  if not user_mac then
-    portal_proxy.no_dhcp_lease()
-    os.exit()
-  end
-
-  -- REAUTH CODE BEGIN
-  local connected = portal_proxy.has_user_been_connected(user_mac)
-  if connected == nil then
-    nixio.syslog("err","Request /reauth has failed")
-    return false
-  end
-  if not connected then
-    nixio.syslog("err","Failed to get user status from db")
-    return false
-  end
-  if connected["authenticated"] then
-    -- If user has been pre authenticated on this AP but then has authenticated on another AP
-    if status == "User in localdb" then
-      local find = "/usr/bin/find /var/localdb/%s -name '*' -type d -mindepth 3"
-      find = string.format(find,user_mac)
-      local path = io.popen(find):read("*l")
-      local remove = "/bin/rm -rf /var/localdb/" .. user_mac
-      s = os.execute(remove)
-      if s ~= 0 then
-        nixio.syslog("err","Failed to remove /var/localdb/" .. mac)
-        return false
-      end
+  
+  local user_status = portal_proxy.status_user(user_ip,user_mac)
+  if user_status ~= 'Authenticated' then
+    -- REAUTH CODE BEGIN
+    local connected = portal_proxy.has_user_been_connected(user_mac)
+    if connected == nil then
+      nixio.syslog("err","Request /reauth has failed")
+      return false
     end
-    local sid = connected["sessionid"]
-    local secret = connected["secret"]
-    local user_id = connected["user_id"]
-    local date_auth = tonumber(connected["ap_validation"])
-    portal_proxy.reauthenticate_user(user_ip,user_mac,sid,secret,date_auth,user_id)
-    uhttpd.send('Status: 200 OK\r\n')
-    uhttpd.send('Content-Type: text/html\r\n\r\n')
-    uhttpd.send('{"url":"' .. cst.landing_page .. '"}')
-    return true
+    if not connected then
+      nixio.syslog("err","Failed to get user status from db")
+      return false
+    end
+    if connected["authenticated"] then
+      -- If user has been pre authenticated on this AP but then has authenticated on another AP
+      if status == "User in localdb" then
+        local find = "/usr/bin/find /var/localdb/%s -name '*' -type d -mindepth 3"
+        find = string.format(find,user_mac)
+        local path = io.popen(find):read("*l")
+        local remove = "/bin/rm -rf /var/localdb/" .. user_mac
+        s = os.execute(remove)
+        if s ~= 0 then
+          nixio.syslog("err","Failed to remove /var/localdb/" .. mac)
+          return false
+        end
+      end
+      local sid = connected["sessionid"]
+      local secret = connected["secret"]
+      local user_id = connected["user_id"]
+      local date_auth = tonumber(connected["ap_validation"])
+      portal_proxy.reauthenticate_user(user_ip,user_mac,sid,secret,date_auth,user_id)
+      uhttpd.send('Status: 200 OK\r\n')
+      uhttpd.send('Content-Type: text/html\r\n\r\n')
+      uhttpd.send('{"url":"' .. cst.landing_page .. '"}')
+      return true
+    end
   end
   -- REAUTH CODE END
